fix last_operation by multiplexing the operation string
As last_operation does not require the service_id parameter, we have to
improvise in order to map the call to the right Service. We do this by
prepending the service_id to the operation string, and parsing it back.

The only drawback (aside from being a workaround) is that when upgrading
from a previous version of this service broker, there can not be any
async requests in flight.